dhcpv4: move all list of IPs types to options_ips.go.
